move class and function to seperate file
