+++ app/src/main/java/com/herlianzhang/mikropos/App.kt
@@ -1,8 +1,6 @@
 package com.herlianzhang.mikropos
 
 import android.app.Application
-import com.google.firebase.crashlytics.ktx.crashlytics
-import com.google.firebase.ktx.Firebase
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
