+++ app/src/main/java/com/herlianzhang/mikropos/ui/customer/createcustomer/CreateCustomerEvent.kt
@@ -1,0 +1,6 @@
+package com.herlianzhang.mikropos.ui.customer.createcustomer
+
+sealed class CreateCustomerEvent {
+    data class ShowErrorSnackbar(val message: String?) : CreateCustomerEvent()
+    object BackWithResult : CreateCustomerEvent()
+}
