+++ app/src/main/java/com/herlianzhang/mikropos/ui/customer/customerdetail/CustomerDetailEvent.kt
@@ -1,0 +1,8 @@
+package com.herlianzhang.mikropos.ui.customer.customerdetail
+
+sealed class CustomerDetailEvent {
+    data class ShowErrorSnackbar(val message: String?) : CustomerDetailEvent()
+    object HideDialog : CustomerDetailEvent()
+    object SetHasChanges : CustomerDetailEvent()
+    object Back : CustomerDetailEvent()
+}
