+++ app/src/main/java/com/herlianzhang/mikropos/ui/home/HomeEvent.kt
@@ -1,0 +1,5 @@
+package com.herlianzhang.mikropos.ui.home
+
+sealed class HomeEvent {
+    object Logout: HomeEvent()
+}
