+++ app/src/main/java/com/herlianzhang/mikropos/ui/home/HomeScreen.kt
@@ -4,13 +4,10 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.rounded.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
@@ -20,50 +17,11 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import com.herlianzhang.mikropos.ui.common.Screen
 import com.herlianzhang.mikropos.ui.setting.MenuScreen
 import com.herlianzhang.mikropos.ui.transaction.cart.CartScreen
 import com.herlianzhang.mikropos.ui.transaction.transactionlist.TransactionListScreen
 import kotlinx.coroutines.flow.collectLatest
-
-sealed class Screen(val route: String, val label: String, val icon: ImageVector) {
-    object Cart : Screen("cart", "Keranjang", Icons.Rounded.AddShoppingCart)
-    object TransactionList : Screen("transactionList", "Transaksi", Icons.Rounded.ReceiptLong)
-    object Menu : Screen("menu", "Menu", Icons.Rounded.Menu)
-}
-
-fun getLabel(route: String?): String {
-    return when (route) {
-        Screen.Cart.route -> Screen.Cart.label
-        Screen.TransactionList.route -> Screen.TransactionList.label
-        Screen.Menu.route -> Screen.Menu.label
-        else -> ""
-    }
-}
-
-@Composable
-fun HomeAction(route: String?, viewModel: HomeViewModel) {
-    when (route) {
-        Screen.Cart.route -> {
-            IconButton(onClick = {
-                throw RuntimeException("percobaan ke dua")
-            }) {
-                Icon(Icons.Rounded.Add, contentDescription = null)
-            }
-        }
-        Screen.TransactionList.route -> {
-            IconButton(onClick = {}) {
-                Icon(Icons.Rounded.QrCodeScanner, contentDescription = null)
-            }
-        }
-        Screen.Menu.route -> {
-            IconButton(onClick = {
-                viewModel.logout()
-            }) {
-                Icon(Icons.Rounded.Logout, contentDescription = null)
-            }
-        }
-    }
-}
 
 @Composable
 fun HomeScreen(
@@ -94,7 +52,7 @@ fun HomeScreen(
             TopAppBar {
                 Spacer(Modifier.width(48.dp))
                 Text(
-                    getLabel(currentRoute),
+                    Screen.getLabel(currentRoute),
                     modifier = Modifier.weight(1f),
                     textAlign = TextAlign.Center
                 )
