+++ app/src/main/java/com/herlianzhang/mikropos/ui/login/LoginEvent.kt
@@ -1,0 +1,6 @@
+package com.herlianzhang.mikropos.ui.login
+
+sealed class LoginEvent {
+    data class ShowErrorSnackbar(val message: String?) : LoginEvent()
+    object NavigateToHome : LoginEvent()
+}
