+++ app/src/main/java/com/herlianzhang/mikropos/ui/login/LoginScreen.kt
@@ -43,7 +43,7 @@ fun LoginScreen(
     val scaffoldState = rememberScaffoldState()
     val isLoading by viewModel.isLoading.collectAsState()
 
-    LaunchedEffect(Unit) {
+    LaunchedEffect(true) {
         viewModel.event.collectLatest { event ->
             when (event) {
                 is LoginEvent.ShowErrorSnackbar -> {
