+++ app/src/main/java/com/herlianzhang/mikropos/ui/login/LoginViewModel.kt
@@ -7,15 +7,12 @@ import com.herlianzhang.mikropos.repository.UserRepository
 import com.herlianzhang.mikropos.utils.UserPreferences
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
-import timber.log.Timber
 import javax.inject.Inject
-
-sealed class LoginEvent {
-    data class ShowErrorSnackbar(val message: String?) : LoginEvent()
-    object NavigateToHome : LoginEvent()
-}
 
 @HiltViewModel
 class LoginViewModel @Inject constructor(
