+++ app/src/main/java/com/herlianzhang/mikropos/ui/printer/PrinterListScreen.kt
@@ -5,11 +5,9 @@ import android.bluetooth.BluetoothAdapter
 import android.content.Intent
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.rounded.ArrowBack
@@ -21,9 +19,6 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.herlianzhang.mikropos.ui.common.LoadingView
@@ -82,43 +77,8 @@ fun PrinterListScreen(viewModel: PrinterListViewModel) {
                     verticalArrangement = Arrangement.spacedBy(16.dp)
                 ) {
                     items(printers) { item ->
-                        Row(
-                            modifier = Modifier
-                                .clip(RoundedCornerShape(12.dp))
-                                .border(1.dp, if (item.isConnected) MaterialTheme.colors.primarySurface else Color.Gray, RoundedCornerShape(12.dp))
-                                .padding(16.dp),
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Column(
-                                modifier = Modifier.weight(1f),
-                                verticalArrangement = Arrangement.spacedBy(12.dp)
-                            ) {
-                                Text(
-                                    item.device.name ?: "-",
-                                    style = MaterialTheme.typography
-                                        .h6
-                                    .copy(fontWeight = FontWeight.Bold)
-                                )
-                                Text(
-                                    item.device.address ?: "-",
-                                    style = MaterialTheme
-                                        .typography
-                                        .body2
-                                        .copy(color = Color.Gray)
-                                )
-                            }
-                            if (!item.isConnected) {
-                                TextButton(onClick = {
-                                    viewModel.connect(item.device.address ?: null)
-                                }) {
-                                    Text("Hubungkan")
-                                }
-                            } else {
-                                Text(
-                                    "Terhubung",
-                                    color = Color.Green
-                                )
-                            }
+                        PrinterItem(item) {
+                            viewModel.connect(item.device.address ?: null)
                         }
                     }
                 }
