+++ app/src/main/java/com/herlianzhang/mikropos/ui/product/createproduct/CreateProductEvent.kt
@@ -1,0 +1,6 @@
+package com.herlianzhang.mikropos.ui.product.createproduct
+
+sealed class CreateProductEvent {
+    data class ShowErrorSnackbar(val message: String?) : CreateProductEvent()
+    object BackWithResult : CreateProductEvent()
+}
