+++ app/src/main/java/com/herlianzhang/mikropos/ui/product/createproduct/CreateProductViewModel.kt
@@ -25,11 +25,6 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
-sealed class CreateProductEvent {
-    data class ShowErrorSnackbar(val message: String?) : CreateProductEvent()
-    object BackWithResult : CreateProductEvent()
-}
-
 @HiltViewModel
 class CreateProductViewModel @Inject constructor(
     private val imageRepository: ImageRepository,
