+++ app/src/main/java/com/herlianzhang/mikropos/ui/product/productdetail/ProductDetailEvent.kt
@@ -1,0 +1,8 @@
+package com.herlianzhang.mikropos.ui.product.productdetail
+
+sealed class ProductDetailEvent {
+    data class ShowErrorSnackbar(val message: String?) : ProductDetailEvent()
+    object HideDialog : ProductDetailEvent()
+    object SetHasChanges : ProductDetailEvent()
+    object Back : ProductDetailEvent()
+}
