+++ app/src/main/java/com/herlianzhang/mikropos/ui/register/RegisterViewModel.kt
@@ -7,14 +7,12 @@ import com.herlianzhang.mikropos.repository.UserRepository
 import com.herlianzhang.mikropos.utils.UserPreferences
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.flow.*
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
-
-sealed class RegisterEvent {
-    data class ShowErrorSnackbar(val message: String?) : RegisterEvent()
-    object NavigateToHome : RegisterEvent()
-}
 
 @HiltViewModel
 class RegisterViewModel @Inject constructor(
