+++ app/src/main/java/com/herlianzhang/mikropos/ui/stock/stocklist/StockListEvent.kt
@@ -1,0 +1,5 @@
+package com.herlianzhang.mikropos.ui.stock.stocklist
+
+sealed class StockListEvent {
+    data class ShowErrorSnackbar(val message: String?) : StockListEvent()
+}
