+++ app/src/main/java/com/herlianzhang/mikropos/ui/theme/Theme.kt
@@ -2,7 +2,6 @@ package com.herlianzhang.mikropos.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
-import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
