+++ app/src/main/java/com/herlianzhang/mikropos/utils/extensions/Int+Extensions.kt
@@ -1,7 +1,6 @@
 package com.herlianzhang.mikropos.utils.extensions
 
 import java.text.NumberFormat
-import java.text.SimpleDateFormat
 import java.util.*
 
 fun Int?.toRupiah(): String {
